sentinel_data/requirements.txt
sqlalchemy
asyncpg
alembic
aioboto3
qdrant-client
nats-py
pydantic-settings
psycopg2-binary
